# apps/web/public/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Dashwise API",
		"version": "1.0.0",
		"description": "OpenAPI specification for Dashwise v1 API routes."
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	],
	"tags": [
		{ "name": "app" },
		{ "name": "auth" },
		{ "name": "config" },
		{ "name": "integrations" },
		{ "name": "jobs" },
		{ "name": "links" },
		{ "name": "locations" },
		{ "name": "monitoring" },
		{ "name": "news" },
		{ "name": "notifications" },
		{ "name": "pageConfig" },
		{ "name": "search" },
		{ "name": "test" },
		{ "name": "wallpapers" },
		{ "name": "widgets" },
		{ "name": "weather" }
	],
	"paths": {
		"/appConfig": {
			"get": {
				"tags": ["app"],
				"summary": "Get application config",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/appInfo": {
			"get": {
				"tags": ["app"],
				"summary": "Get application info",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/auth/callback": {
			"get": {
				"tags": ["auth"],
				"summary": "Auth callback",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"302": { "description": "Redirect" }
				}
			}
		},
		"/auth/change-password": {
			"post": {
				"tags": ["auth"],
				"summary": "Change password",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/auth/delete-account": {
			"delete": {
				"tags": ["auth"],
				"summary": "Delete account",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/auth/login": {
			"post": {
				"tags": ["auth"],
				"summary": "Login",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/auth/mfa": {
			"post": {
				"tags": ["auth"],
				"summary": "MFA verification",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/auth/signup": {
			"post": {
				"tags": ["auth"],
				"summary": "Signup",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"400": { "$ref": "#/components/responses/JsonBadRequest" }
				}
			}
		},
		"/auth/sso": {
			"get": {
				"tags": ["auth"],
				"summary": "SSO start",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"302": { "description": "Redirect" }
				}
			}
		},
		"/auth/validate-auth": {
			"post": {
				"tags": ["auth"],
				"summary": "Validate auth",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/auth/update-user-property": {
			"patch": {
				"tags": ["auth"],
				"summary": "Update user property",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/config": {
			"get": {
				"tags": ["config"],
				"summary": "Get config",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["config"],
				"summary": "Create config",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"patch": {
				"tags": ["config"],
				"summary": "Patch config",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"put": {
				"tags": ["config"],
				"summary": "Replace config",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/config/delete-unused-linkgroups": {
			"post": {
				"tags": ["config"],
				"summary": "Delete unused link groups",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/config/migrate-legacy-page-config": {
			"post": {
				"tags": ["config"],
				"summary": "Migrate legacy page config",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/config/move-arrayitems": {
			"post": {
				"tags": ["config"],
				"summary": "Move config array items",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/beszel": {
			"get": {
				"tags": ["integrations"],
				"summary": "Get Beszel integration",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/beszel/system-healthstats": {
			"get": {
				"tags": ["integrations"],
				"summary": "Get Beszel system health stats",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/dashdot": {
			"post": {
				"tags": ["integrations"],
				"summary": "Proxy Dashdot integration",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/karakeep": {
			"get": {
				"tags": ["integrations"],
				"summary": "Get Karakeep integration",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations": {
			"get": {
				"tags": ["integrations"],
				"summary": "List integrations",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["integrations"],
				"summary": "Create integration",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/test-endpoint": {
			"post": {
				"tags": ["integrations"],
				"summary": "Test integration endpoint",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/widget-properties": {
			"get": {
				"tags": ["integrations"],
				"summary": "Get widget properties",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/integrations/with-widget": {
			"get": {
				"tags": ["integrations"],
				"summary": "Get integration with widget",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/collections": {
			"get": {
				"tags": ["links"],
				"summary": "Get link collections",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/home/groups": {
			"get": {
				"tags": ["links"],
				"summary": "Get home link groups",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["links"],
				"summary": "Create home link group",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/home": {
			"get": {
				"tags": ["links"],
				"summary": "Get home links",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/folders": {
			"get": {
				"tags": ["links"],
				"summary": "Get link folders",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/items": {
			"get": {
				"tags": ["links"],
				"summary": "Get link items",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["links"],
				"summary": "Create link item",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/items/{linkId}": {
			"put": {
				"tags": ["links"],
				"summary": "Update link item",
				"parameters": [
					{
						"name": "linkId",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"delete": {
				"tags": ["links"],
				"summary": "Delete link item",
				"parameters": [
					{
						"name": "linkId",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/links/tags": {
			"get": {
				"tags": ["links"],
				"summary": "Get link tags",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/jobs/searchItems": {
			"get": {
				"tags": ["jobs"],
				"summary": "Search items job",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/jobs/pullIcons": {
			"get": {
				"tags": ["jobs"],
				"summary": "Pull icons job",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"409": { "$ref": "#/components/responses/JsonBadRequest" }
				}
			}
		},
		"/locations": {
			"get": {
				"tags": ["locations"],
				"summary": "Get locations",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/monitoringStatus": {
			"get": {
				"tags": ["monitoring"],
				"summary": "Get monitoring status",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["monitoring"],
				"summary": "Trigger monitoring check for one item",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"linkId": { "type": "string" },
									"jobId": { "type": "string" }
								},
								"additionalProperties": false
							}
						}
					}
				},
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" },
					"400": { "$ref": "#/components/responses/JsonBadRequest" },
					"401": { "$ref": "#/components/responses/JsonUnauthorized" }
				}
			}
		},
		"/news": {
			"get": {
				"tags": ["news"],
				"summary": "Get news subscriptions",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed": {
			"get": {
				"tags": ["news"],
				"summary": "Get news feed",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed-category-rename": {
			"post": {
				"tags": ["news"],
				"summary": "Rename feed category",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed-refresh": {
			"post": {
				"tags": ["news"],
				"summary": "Refresh feeds",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"get": {
				"tags": ["news"],
				"summary": "Get feed refresh status",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed-subscribe": {
			"post": {
				"tags": ["news"],
				"summary": "Subscribe to feed",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed-unsubscribe": {
			"post": {
				"tags": ["news"],
				"summary": "Unsubscribe from feed",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/news/feed-update": {
			"post": {
				"tags": ["news"],
				"summary": "Update feed",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications": {
			"get": {
				"tags": ["notifications"],
				"summary": "Get notifications",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["notifications"],
				"summary": "Create notification",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications/forwarders": {
			"get": {
				"tags": ["notifications"],
				"summary": "List forwarders",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["notifications"],
				"summary": "Create forwarder",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"put": {
				"tags": ["notifications"],
				"summary": "Update forwarder",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"delete": {
				"tags": ["notifications"],
				"summary": "Delete forwarder",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications/markAsRead": {
			"post": {
				"tags": ["notifications"],
				"summary": "Mark notifications as read",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications/{topic}": {
			"post": {
				"tags": ["notifications"],
				"summary": "Send notification to topic",
				"parameters": [
					{
						"name": "topic",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications/topics": {
			"get": {
				"tags": ["notifications"],
				"summary": "List topics",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"post": {
				"tags": ["notifications"],
				"summary": "Create topic",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/notifications/topicTokens": {
			"post": {
				"tags": ["notifications"],
				"summary": "Create topic token",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"get": {
				"tags": ["notifications"],
				"summary": "List topic tokens",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"delete": {
				"tags": ["notifications"],
				"summary": "Delete topic token",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/searchItems": {
			"get": {
				"tags": ["search"],
				"summary": "Search items",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/pageConfig": {
			"get": {
				"tags": ["pageConfig"],
				"summary": "Get page config",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"put": {
				"tags": ["pageConfig"],
				"summary": "Update page config",
				"requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/pageConfig/user-pages": {
			"get": {
				"tags": ["pageConfig"],
				"summary": "Get user pages",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/test/bookmarks": {
			"get": {
				"tags": ["test"],
				"summary": "Test bookmarks",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/wallpapers": {
			"post": {
				"tags": ["wallpapers"],
				"summary": "Upload or update wallpaper",
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									},
									"fileName": {
										"type": "string",
										"description": "Optional fallback file name"
									},
									"convertToWebp": {
										"type": "boolean",
										"description": "Convert the uploaded image to WebP before storing"
									}
								},
								"required": ["image"]
							}
						}
					}
				},
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			},
			"get": {
				"tags": ["wallpapers"],
				"summary": "Get wallpaper",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/widgets": {
			"get": {
				"tags": ["widgets"],
				"summary": "Get user widgets",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/widgets/glanceable": {
			"get": {
				"tags": ["widgets"],
				"summary": "Get user glanceable",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/widgets/glanceables": {
			"get": {
				"tags": ["widgets"],
				"summary": "Get user glanceables",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		},
		"/weather": {
			"get": {
				"tags": ["weather"],
				"summary": "Get weather",
				"responses": {
					"200": { "$ref": "#/components/responses/JsonOk" }
				}
			}
		}
	},
	"components": {
		"schemas": {
			"GenericObject": {
				"type": "object",
				"additionalProperties": true
			},
			"Error": {
				"type": "object",
				"additionalProperties": true
			}
		},
		"requestBodies": {
			"JsonBody": {
				"required": false,
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/GenericObject" }
					}
				}
			}
		},
		"responses": {
			"JsonOk": {
				"description": "OK",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/GenericObject" }
					}
				}
			},
			"JsonBadRequest": {
				"description": "Bad Request",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" }
					}
				}
			},
			"JsonUnauthorized": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": { "$ref": "#/components/schemas/Error" }
					}
				}
			}
		}
	}
}